requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.2.0
boto==2.49.0
Brotli==1.0.9
cffi==1.15.0
chardet==4.0.0
discord.py==2.2.2
emoji==1.6.1
idna==3.3
multidict==5.2.0
numpy==1.21.4
orjson==3.6.5
pandas==1.3.5
pycares==4.1.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML~=6.0.2
requests==2.28.1
six==1.16.0
typing_extensions==4.0.1
xlrd==2.0.1
yarl==1.7.2
z3==0.2.0
